# Replacing costly merges in UGraph.