@SkySkimmer 2017-04-18
Could you bench the result of using Map.union (needs ocaml > 4.03)? I'm curious. (or tell me how to do coq benchmarks)