@ppedrot 2017-04-18
You can go to https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/build if you have an account on ci.inria.fr in the Coq project, and fill the missing data. If you don't have such an account I think I am an administrator so that I can add you if you open a account first.

Note that I'm pretty sure it is going to be slower as Map.union is more or less a specialized version of Map.merge, and thus that has to crawl both arguments.