@ppedrot 2017-04-18
I've just realized that the bench claims to have failed on this run, but I can't find where in the logs. @matejkosik Any idea on how to extract the relevant information [from them](https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/75/consoleText)?