@JasonGross 2017-04-19
@ppedrot Searching for "Error 1" in the log gives:
```
- File "./Tarski_dev/Ch16_coordinates_with_functions.v", line 1314, characters 11-12:
- Error: The reference H was not found in the current environment.
- could not prove discrimination result
- could not prove discrimination result
- could not prove discrimination result
- could not prove discrimination result
- could not prove discrimination result
- could not prove discrimination result
- could not prove discrimination result
- could not prove discrimination result
- Makefile:418: recipe for target 'Tarski_dev/Ch16_coordinates_with_functions.vo' failed
- 
- make: *** [Tarski_dev/Ch16_coordinates_with_functions.vo] Error 1
[ERROR] The compilation of coq-geocoq failed at "make -j1".
```
and
```
#=== ERROR while installing coq-geocoq.dev ====================================#
# opam-version         1.2.2
# os                   linux
# command              make -j1
# path                 /home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-geocoq.dev
# compiler             4.04.0
# exit-code            2
# env-file             /home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-geocoq.dev/coq-geocoq-32635-e7966f.env
# stdout-file          /home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-geocoq.dev/coq-geocoq-32635-e7966f.out
# stderr-file          /home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-geocoq.dev/coq-geocoq-32635-e7966f.err
### stdout ###
# [...]
# could not prove discrimination result
# could not prove discrimination result
# could not prove discrimination result
# could not prove discrimination result
# could not prove discrimination result
# could not prove discrimination result
# could not prove discrimination result
# could not prove discrimination result
# could not prove discrimination result
# Makefile:418: recipe for target 'Tarski_dev/Ch16_coordinates_with_functions.vo' failed
### stderr ###
# File "./Tarski_dev/Ch16_coordinates_with_functions.v", line 5, characters 0-22:
# Warning: Notation [ _ ] was already used [notation-overridden,parsing]
# File "./Tarski_dev/Ch16_coordinates_with_functions.v", line 1314, characters 11-12:
# Error: The reference H was not found in the current environment.
# 
# make: *** [Tarski_dev/Ch16_coordinates_with_functions.vo] Error 1
```
and
```
- File "./finmap.v", line 2297, characters 0-72:
- Error:
- Ltac call to "by (ssrhintarg)" failed.
- Conversion test raised an anomaly
- 
- Makefile.coq:271: recipe for target 'finmap.vo' failed
- make[1]: Leaving directory '/home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-mathcomp-finmap.dev'
- Makefile:4: recipe for target 'all' failed
- make[1]: *** [finmap.vo] Error 1
- make: *** [all] Error 2
[ERROR] The compilation of coq-mathcomp-finmap failed at "make -j 1".
```
and
```
#=== ERROR while installing coq-mathcomp-finmap.dev ===========================#
# opam-version         1.2.2
# os                   linux
# command              make -j 1
# path                 /home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-mathcomp-finmap.dev
# compiler             4.04.0
# exit-code            2
# env-file             /home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-mathcomp-finmap.dev/coq-mathcomp-finmap-15927-6d07ae.env
# stdout-file          /home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-mathcomp-finmap.dev/coq-mathcomp-finmap-15927-6d07ae.out
# stderr-file          /home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-mathcomp-finmap.dev/coq-mathcomp-finmap-15927-6d07ae.err
### stdout ###
# [...]
# make -f Makefile.coq
# make[1]: Entering directory '/home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-mathcomp-finmap.dev'
# COQDEP set.v
# COQDEP order.v
# COQDEP multiset.v
# COQDEP finmap.v
# COQC finmap.v
# Makefile.coq:271: recipe for target 'finmap.vo' failed
# make[1]: Leaving directory '/home/jenkins/workspace/benchmark-part-of-the-branch/75/.opam.NEW/4.04.0/build/coq-mathcomp-finmap.dev'
# Makefile:4: recipe for target 'all' failed
### stderr ###
# [...]
# File "./finmap.v", line 1867, characters 0-49:
# Warning: Implicit Arguments is deprecated; use Arguments instead
# [deprecated-implicit-arguments,deprecated]
# File "./finmap.v", line 2297, characters 0-72:
# Error:
# Ltac call to "by (ssrhintarg)" failed.
# Conversion test raised an anomaly
# 
# make[1]: *** [finmap.vo] Error 1
# make: *** [all] Error 2
```