@silene 2017-04-19
> And now that I'm thinking of it, it's not a union anyway.

How so? It certainly looks like `union (||) to_merge_lvl n.ltle`. This also means that you could make your code a bit more cache-friendly by merging the first and last branch.

> Note that I'm pretty sure it is going to be slower as Map.union is more or less a specialized version of Map.merge, and thus that has to crawl both arguments.

You are presumably right about the slowness, but that is not because of the crawl of both arguments. Union does not scan more nodes than your version, but your version does benefit from an optimization of `add`: it does not create a fresh new node if it would be identical to the old one.