@ppedrot 2017-04-19
@silene All right, I'm going to bench it by curiosity, although it can't go in current trunk.

@JasonGross That's strange, what does this have to do with the patch? It looks unrelated.