@mattam82 2017-04-19
The failures are a bit scary, to me the code looks semantics-preserving but there might be a subtle difference?