@SkySkimmer 2017-04-19
Nevermind the Map.union, UMap is implemented as a hashmap (lib/hMap.ml).