@JasonGross 2017-04-19
@ppedrot Is your question directed at me?  Because I have no idea; you just asked for a way to extract the errors from the log you linked to, and that's what I was giving you