@ppedrot 2017-04-19
@mattam82 It seems to be already the case that geo-coq is broken before this patch. So something happened in between.