@matejkosik 2017-04-20
@ppedrot I need to load an updated version of benchmarking scripts to `pendulum` (with less debugging outputs). I'll try to do that at a point when no benchmarking will be in progress. The debugging output dillutes the relevant information printed by the benchmarking script, which is:
```
INFO: the following OPAM-packages:
- coq-geocoq
- coq-mathcomp-finmap
are not installable.
```
This is something we already [know](https://ci.inria.fr/coq/view/opam/job/opam-install-trunk/153/console) and ignore (?).