@ppedrot 2017-04-30
@mattam82 Just for the sake of having a formal approval, do you confirm the code is correct for you? That's a stupid optimization, the sooner we merge it the better.