@ppedrot 2017-05-10
@maximedenes I think you may review this patch yourself if we don't manage to get @mattam82's attention. It's really a trivial, purely local argument on the semantic of `Map.merge` and does not require any understanding of universes at all.